feature/86cv9zk19 rename utility class, add comments
