+++ behind-the-veil-siteroot/imports/ui/components/map/mapUtils.js
